requirements.txt
django
django-test-extensions
textile
django-markup
django-tagging
django-comments-spamfighter
mysql-python
django-nose
pil
django-push
